// api-scoring-extension/plugins/spa-apiscoringviewer/packages/apiscoringviewer/src/types.ts
export type ProtocolType = "REST" | "EVENT" | "GRPC";
export type Rating = "A+" | "A" | "B" | "C" | "D";
export type ValidationType = "LINTER" | "DESIGN" | "DOCUMENTATION" | "SECURITY" | "OVERALL_SCORE";

export enum Severity {
  Error = 0,
  Warning,
  Info,
}

export type VSCodeCommand<C extends string, P> = { command: C; payload: P };

export type SetCertificationResults<TApiIdentifier extends ApiIdentifier> = {
  command: "setCertificationResults";
  payload: CertificationPayload<TApiIdentifier>;
};

export type SetModuleResults = {
  command: "setModuleResults";
  payload: ModulePayload;
};

export type SetFileResults = {
  command: "setFileResults";
  payload: FilePayload;
};

export type SetFileResultsError = VSCodeCommand<"setFileResultsError", Error>;

export interface MessageHandler<P> {
  (payload: P): void;
}

export type CertificationPayload<TApiIdentifier extends ApiIdentifier> = {
  metadata: {
    apis: { name: string; apiSpecType: ProtocolType; definitionPath: string }[];
  };
  results: Certification<TApiIdentifier>[];
  rootPath?: string;
};

export type ModulePayload<TApiIdentifier extends ApiIdentifier = ApiIdentifier> = {
  apiModule: ModuleValidation;
  results: Certification<TApiIdentifier>[];
};

export type ValidationResponse<TApiIdentifier extends ApiIdentifier = ApiIdentifier> = Certification<TApiIdentifier>;

export type GetApiIdentifier<TApiIdentifier extends ApiIdentifier> = (apiCertification: Certification<TApiIdentifier>) => string;

export type ApiIdentifier = {
  apiName: string;
  apiProtocol: ProtocolType;
};

export type Certification<TApiIdentifier extends ApiIdentifier = ApiIdentifier> = TApiIdentifier & {
  rating?: Rating;
  ratingDescription: string;
  score: number;
  result: ({ designValidation: CodeValidation } | { securityValidation: CodeValidation } | { documentationValidation: DocValidation })[];
  apiType?: string;
};
// ...
export type DataProviderChildFn<TApiIdentifier extends ApiIdentifier = ApiIdentifier> = (props: {
  certification?: CertificationPayload<TApiIdentifier> | null;
  loading: boolean;
  error?: Error | null;
  modulesMetadata: ModulesMetadata;
  apisRevalidationMetadata: ModulesMetadata;
  revalidateModule?: (validationBody: ModuleValidation) => void;
  revalidateApi?: (validationBody: ModuleValidation) => void;
}) => JSX.Element;

export type ModuleValidation = {
  apiName: string;
  apiDefinitionPath?: string;
  validationType: ValidationType;
  apiSpecType: ProtocolType;
};

export type ModuleMetadata = { loading: boolean };
export type ModulesMetadata = Record<string, ModuleMetadata>;